requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
colorama>=0.4.4
faiss-cpu>=1.7.4
numpy>=1.21.0
requests>=2.26.0
beautifulsoup4>=4.9.3
lxml>=4.9.0 